requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export -o requirements.txt --no-hashes
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.17
    # via gcsfs
aiosignal==1.3.2
    # via aiohttp
altair==5.5.0
    # via streamlit
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
beautifulsoup4==4.13.4
    # via ocr-test
blinker==1.9.0
    # via streamlit
boto3==1.37.28
    # via ocr-test
botocore==1.37.28
    # via
    #   boto3
    #   s3transfer
bottleneck==1.4.2
    # via pandas
cachetools==5.5.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.1.31
    # via requests
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.1
    # via
    #   pdfminer-six
    #   requests
click==8.1.8
    # via streamlit
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
cryptography==44.0.2
    # via pdfminer-six
db-dtypes==1.4.2
    # via pandas-gbq
decorator==5.2.1
    # via gcsfs
deprecated==1.2.18
    # via pikepdf
deprecation==2.1.0
    # via ocrmypdf
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.3.2
    # via gcsfs
gcsfs==2025.3.2
    # via pandas
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
google-api-core==2.24.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-documentai
    #   google-cloud-documentai-toolbox
    #   google-cloud-storage
    #   google-cloud-vision
    #   pandas-gbq
google-auth==2.39.0
    # via
    #   gcsfs
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-documentai
    #   google-cloud-storage
    #   google-cloud-vision
    #   pandas-gbq
    #   pydata-google-auth
google-auth-oauthlib==1.2.1
    # via
    #   gcsfs
    #   pandas-gbq
    #   pydata-google-auth
google-cloud-bigquery==3.31.0
    # via
    #   google-cloud-documentai-toolbox
    #   pandas-gbq
google-cloud-core==2.4.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-documentai==3.3.0
    # via google-cloud-documentai-toolbox
google-cloud-documentai-toolbox==0.14.2a0
    # via ocr-test
google-cloud-storage==3.1.0
    # via
    #   gcsfs
    #   google-cloud-documentai-toolbox
google-cloud-vision==3.10.1
    # via google-cloud-documentai-toolbox
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-documentai-toolbox
grpcio==1.71.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.71.0
    # via google-api-core
idna==3.10
    # via
    #   requests
    #   yarl
img2pdf==0.6.0
    # via ocrmypdf
immutabledict==4.2.1
    # via google-cloud-documentai-toolbox
intervaltree==3.1.0
    # via google-cloud-documentai-toolbox
jinja2==3.1.6
    # via
    #   altair
    #   google-cloud-documentai-toolbox
    #   pydeck
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
llvmlite==0.44.0
    # via numba
lxml==5.3.2
    # via
    #   ocr-test
    #   pikepdf
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
narwhals==1.33.0
    # via altair
numba==0.61.2
    # via pandas
numexpr==2.10.2
    # via pandas
numpy==1.26.4
    # via
    #   bottleneck
    #   db-dtypes
    #   google-cloud-documentai-toolbox
    #   numba
    #   numexpr
    #   pandas
    #   pandas-gbq
    #   pyarrow
    #   pydeck
    #   streamlit
oauthlib==3.2.2
    # via requests-oauthlib
ocrmypdf==16.10.1
    # via ocr-test
packaging==24.2
    # via
    #   altair
    #   db-dtypes
    #   deprecation
    #   google-cloud-bigquery
    #   ocrmypdf
    #   pandas-gbq
    #   pikepdf
    #   streamlit
pandas==2.2.3
    # via
    #   db-dtypes
    #   google-cloud-documentai-toolbox
    #   pandas-gbq
    #   streamlit
pandas-gbq==0.28.0
    # via pandas
pdfminer-six==20250416
    # via ocrmypdf
pi-heif==0.22.0
    # via ocrmypdf
pikepdf==8.15.1
    # via
    #   google-cloud-documentai-toolbox
    #   img2pdf
    #   ocrmypdf
pillow==10.4.0
    # via
    #   google-cloud-documentai-toolbox
    #   img2pdf
    #   ocrmypdf
    #   pi-heif
    #   pikepdf
    #   streamlit
pluggy==1.5.0
    # via ocrmypdf
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-documentai
    #   google-cloud-documentai-toolbox
    #   google-cloud-vision
protobuf==5.29.4
    # via
    #   google-api-core
    #   google-cloud-documentai
    #   google-cloud-vision
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
    #   streamlit
pyarrow==15.0.2
    # via
    #   db-dtypes
    #   google-cloud-documentai-toolbox
    #   pandas-gbq
    #   streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22 ; platform_python_implementation != 'PyPy'
    # via cffi
pydata-google-auth==1.9.1
    # via pandas-gbq
pydeck==0.9.1
    # via streamlit
pygments==2.19.1
    # via rich
pymupdf==1.25.5
    # via ocr-test
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   google-cloud-bigquery
    #   pandas
pytz==2025.2
    # via pandas
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   gcsfs
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   requests-oauthlib
    #   streamlit
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==14.0.0
    # via ocrmypdf
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
ruff==0.11.6
s3transfer==0.11.4
    # via boto3
setuptools==78.1.1
    # via
    #   pandas-gbq
    #   pydata-google-auth
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sortedcontainers==2.4.0
    # via intervaltree
soupsieve==2.7
    # via beautifulsoup4
streamlit==1.44.1
    # via ocr-test
tabulate==0.9.0
    # via google-cloud-documentai-toolbox
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
typing-extensions==4.13.1
    # via
    #   altair
    #   beautifulsoup4
    #   referencing
    #   streamlit
tzdata==2025.2
    # via pandas
urllib3==2.3.0
    # via
    #   botocore
    #   requests
watchdog==6.0.0 ; sys_platform != 'darwin'
    # via streamlit
wrapt==1.17.2
    # via deprecated
yarl==1.20.0
    # via aiohttp
